DES-177: commit correct component
